requirements.txt
flask
torch
torchvision
torchxrayvision
pydicom
pillow
numpy
captum
requests
